Code-base rearrangement
