+++ vehicle/src/Vehicle/Verify/Core.hs
@@ -1,6 +1,5 @@
 module Vehicle.Verify.Core where
 
-import Control.Monad.IO.Class (MonadIO)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.Map (Map)
 import Data.Map qualified as Map
@@ -9,9 +8,9 @@ import Data.Vector (Vector)
 import Data.Vector qualified as Vector (toList)
 import GHC.Generics (Generic)
 import System.FilePath ((<.>))
+import Vehicle.Backend.Queries.LinearExpr (Assertion, SparseLinearExpr)
+import Vehicle.Backend.Queries.Variable
 import Vehicle.Compile.Prelude (Name)
-import Vehicle.Compile.Queries.LinearExpr (Assertion, CLSTProblem, SparseLinearExpr)
-import Vehicle.Compile.Queries.Variable
 import Vehicle.Compile.Resource
 import Vehicle.Prelude
 
@@ -49,40 +48,6 @@ instance Pretty UserVariableAssignment where
         let name = pretty $ userVarName var
         let valueDoc = prettyConstant True (userVarDimensions var) value
         name <> ":" <+> valueDoc
-
---------------------------------------------------------------------------------
--- Verifiers
-
-data VerifierID
-  = Marabou
-  deriving (Eq, Ord, Show, Read, Bounded, Enum)
-
-instance Pretty VerifierID where
-  pretty = pretty . show
-
--- | Location of the verifier executable file
-type VerifierExecutable = FilePath
-
--- | The type of methods to call a verifier on a query
-type VerifierInvocation =
-  forall m.
-  (MonadIO m) =>
-  VerifierExecutable ->
-  MetaNetwork ->
-  QueryFile ->
-  m (Either Text (QueryResult NetworkVariableAssignment))
-
--- | A complete verifier implementation
-data Verifier = Verifier
-  { -- | The identifier for the verifier within Vehicle itself
-    verifierIdentifier :: VerifierID,
-    -- | The query format that the verifier accepts
-    verifierQueryFormat :: QueryFormatID,
-    -- | The name of the executable for the verifier
-    verifierExecutableName :: String,
-    -- | The command to invoke the verifier
-    invokeVerifier :: VerifierInvocation
-  }
 
 --------------------------------------------------------------------------------
 -- Addresses
@@ -191,27 +156,6 @@ metaNetworkVariables :: Bool -> MetaNetwork -> [NetworkVariable]
 metaNetworkVariables reduced metaNetwork = do
   let (_, result) = foldr (metaNetworkEntryVariables reduced) (mempty, mempty) metaNetwork
   concat (reverse result)
-
---------------------------------------------------------------------------------
--- Query formats
-
-data QueryFormatID
-  = MarabouQueries
-  | VNNLibQueries
-  deriving (Show, Eq, Bounded, Enum)
-
-instance Pretty QueryFormatID where
-  pretty = \case
-    MarabouQueries -> "Marabou query format"
-    VNNLibQueries -> "VNNLib query format"
-
--- | A format for an output query that verifiers can parse.
-data QueryFormat = QueryFormat
-  { queryFormatID :: QueryFormatID,
-    queryOutputFormat :: ExternalOutputFormat,
-    -- | The command to compile an individual query
-    compileQuery :: forall m. (MonadLogger m) => CLSTProblem -> m QueryText
-  }
 
 --------------------------------------------------------------------------------
 -- Query results
